Corrects spelling of SnapshotSerilizer and SnapshotSerializer
