+++ Source/EventFlow/EventFlowOptions.cs
@@ -1,19 +1,19 @@
 // The MIT License (MIT)
-// 
+//
 // Copyright (c) 2015-2021 Rasmus Mikkelsen
 // Copyright (c) 2015-2021 eBay Software Foundation
 // https://github.com/eventflow/EventFlow
-// 
+//
 // Permission is hereby granted, free of charge, to any person obtaining a copy of
 // this software and associated documentation files (the "Software"), to deal in
 // the Software without restriction, including without limitation the rights to
 // use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of
 // the Software, and to permit persons to whom the Software is furnished to do so,
 // subject to the following conditions:
-// 
+//
 // The above copyright notice and this permission notice shall be included in all
 // copies or substantial portions of the Software.
-// 
+//
 // THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 // IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
 // FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
@@ -54,15 +54,17 @@ namespace EventFlow
     public class EventFlowOptions : IEventFlowOptions
     {
         private readonly ConcurrentBag<Type> _aggregateEventTypes = new ConcurrentBag<Type>();
-        private readonly ConcurrentBag<Type> _sagaTypes = new ConcurrentBag<Type>(); 
+        private readonly ConcurrentBag<Type> _sagaTypes = new ConcurrentBag<Type>();
         private readonly ConcurrentBag<Type> _commandTypes = new ConcurrentBag<Type>();
         private readonly EventFlowConfiguration _eventFlowConfiguration = new EventFlowConfiguration();
+
         private readonly ConcurrentBag<Type> _jobTypes = new ConcurrentBag<Type>
             {
                 typeof(PublishCommandJob),
                 typeof(DispatchToAsynchronousEventSubscribersJob)
             };
-        private readonly List<Type> _snapshotTypes = new List<Type>(); 
+
+        private readonly List<Type> _snapshotTypes = new List<Type>();
 
         public IServiceCollection ServiceCollection { get; }
 
@@ -73,7 +75,7 @@ namespace EventFlow
             RegisterDefaults(ServiceCollection);
         }
 
-        public static IEventFlowOptions New() => 
+        public static IEventFlowOptions New() =>
             new EventFlowOptions(
                 new ServiceCollection()
                     .AddLogging(b => b
@@ -170,6 +172,8 @@ namespace EventFlow
         {
             serviceCollection.AddMemoryCache();
 
+            RegisterObsoleteDefaults(serviceCollection);
+
             // Default no-op resilience strategies
             serviceCollection.TryAddTransient<IAggregateStoreResilienceStrategy, NoAggregateStoreResilienceStrategy>();
             serviceCollection.TryAddTransient<IDispatchToReadStoresResilienceStrategy, NoDispatchToReadStoresResilienceStrategy>();
@@ -182,7 +186,7 @@ namespace EventFlow
             serviceCollection.TryAddTransient<ICommandBus, CommandBus>();
             serviceCollection.TryAddTransient<IAggregateStore, AggregateStore>();
             serviceCollection.TryAddTransient<ISnapshotStore, SnapshotStore>();
-            serviceCollection.TryAddTransient<ISnapshotSerilizer, SnapshotSerilizer>();
+            serviceCollection.TryAddTransient<ISnapshotSerializer, SnapshotSerializer>();
             serviceCollection.TryAddTransient<ISnapshotPersistence, NullSnapshotPersistence>();
             serviceCollection.TryAddTransient<ISnapshotUpgradeService, SnapshotUpgradeService>();
             serviceCollection.TryAddTransient<IReadModelPopulator, ReadModelPopulator>();
@@ -223,6 +227,13 @@ namespace EventFlow
                 _aggregateEventTypes,
                 _sagaTypes,
                 _snapshotTypes));
+        }
+
+        private void RegisterObsoleteDefaults(IServiceCollection serviceCollection)
+        {
+#pragma warning disable CS0618 // Type or member is obsolete
+            serviceCollection.TryAddTransient<ISnapshotSerilizer, SnapshotSerilizer>();
+#pragma warning restore CS0618 // Type or member is obsolete
         }
     }
 }
