+++ Source/EventFlow/Snapshots/ISnapshotSerilizer.cs
@@ -1,19 +1,19 @@
 // The MIT License (MIT)
-// 
+//
 // Copyright (c) 2015-2021 Rasmus Mikkelsen
 // Copyright (c) 2015-2021 eBay Software Foundation
 // https://github.com/eventflow/EventFlow
-// 
+//
 // Permission is hereby granted, free of charge, to any person obtaining a copy of
 // this software and associated documentation files (the "Software"), to deal in
 // the Software without restriction, including without limitation the rights to
 // use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of
 // the Software, and to permit persons to whom the Software is furnished to do so,
 // subject to the following conditions:
-// 
+//
 // The above copyright notice and this permission notice shall be included in all
 // copies or substantial portions of the Software.
-// 
+//
 // THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 // IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
 // FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
@@ -21,26 +21,13 @@
 // IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
 // CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
-using System.Threading;
-using System.Threading.Tasks;
-using EventFlow.Core;
+using System;
 
 namespace EventFlow.Snapshots
 {
-    public interface ISnapshotSerilizer
+    [Obsolete("Use 'ISnapshotSerializer' instead")]
+    public interface ISnapshotSerilizer : ISnapshotSerializer
     {
-        Task<SerializedSnapshot> SerializeAsync<TAggregate, TIdentity, TSnapshot>(
-            SnapshotContainer snapshotContainer,
-            CancellationToken cancellationToken)
-            where TAggregate : ISnapshotAggregateRoot<TIdentity, TSnapshot>
-            where TIdentity : IIdentity
-            where TSnapshot : ISnapshot;
-
-        Task<SnapshotContainer> DeserializeAsync<TAggregate, TIdentity, TSnapshot>(
-            CommittedSnapshot committedSnapshot,
-            CancellationToken cancellationToken)
-            where TAggregate : ISnapshotAggregateRoot<TIdentity, TSnapshot>
-            where TIdentity : IIdentity
-            where TSnapshot : ISnapshot;
+        // EMPTY
     }
 }
